feat(v2): lib.rs re-exports v2 types as public surface, Gate 5 commit J
- Make v2 runtime, collection, sorted_collection pub (were pub(crate))
- Remove #![allow(dead_code)] from v2/mod.rs (Gate 5 wires v2 in)
- lib.rs now re-exports from v2: Runtime, Incr, RuntimeId, Value,
  IncrCollection, Delta, GroupedCollection, SortedCollection, SortDelta
- v1 modules kept as private mod items; Task 14 deletes them
- Fix v1 test imports: use crate::runtime::Runtime instead of crate::Runtime
- Fix runtime_vs_v1_bench: use crate::runtime::Runtime and crate::types::Incr
  for v1 aliases so they resolve to v1 module paths, not v2 re-exports
- Update external tests for v2 API: replace Rc<Cell<u32>> with
  Arc<AtomicU32> in create_query closures (v2 requires Send + Sync),
  cast count() result from u64 to usize in property test assertion

+++ crates/incr-core/tests/collection_property.rs
@@ -42,7 +42,7 @@ fn verify_collection_incremental_matches_batch(ops: Vec<Op>) {
         .collect();
 
     assert_eq!(
-        incr_count,
+        incr_count as usize,
         batch_elements.len(),
         "Count mismatch: incr={}, batch={}",
         incr_count,
